Single transfer per client message
This means the server has less work to do keeping track of client
message IDs and correlating them with async responses from ML. This is a
job better performed by the client libraries. Moreover, a large part of
the point of using a websocket transport was to enable interleaving
messages.